Adding support for federated urls in modals
